pyproject.toml
[project]
name = "zania-huaxing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-postgres>=0.0.12",
    "langchain-community>=0.3.0",
    "psycopg[binary,pool]>=3.2.0",
    "chromadb>=0.5.0",
    "pypdf>=5.0.0",
    "streamlit>=1.40.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "ruff>=0.14.4",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=4.0.0",
]
